Modified filenames a bit. (OCD!) Brought everything up to date with previous commit (extensive use of size_t in guts).
git-svn-id: file:///home/jacko/Downloads/LISP_REPO@69 9726ce68-a45a-11dd-9d72-4502ce979ee6

+++ lispy/yalie/guts/array.h
@@ -5,7 +5,7 @@
 
 typedef struct Array * array_t;
 
-array_t new_array( unsigned long int length, void* val );
+array_t new_array( size_t length, void* val );
 void free_array( array_t array );
 
 size_t array_len( array_t array );
